requirements.txt
# requirements.txt
praw>=7.6.0
openai>=0.27.0
python-dotenv>=1.0.0
tiktoken>=0.3.0
